Improve QWizard visuals for macOS dark mode
This did not account for macOS dark mode, with hardcoded white,
but now it correctly updates the colors from palette whenever
the theme is changed.

Additionally changed some bool checks to make sure the code
runs on macOS theme updates.

Do note that this change affects Qt Maintenance Tool (in a
good way).

Pick-to: 5.12 5.15
Change-Id: I6fb92d3f6c15a7ff83d0d8a30cf82477b6aba126
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Katja Marttila <katja.marttila@qt.io>