CMake: Fix building examples targeting iOS simulator with qmake
The makefile generated by qmake when building an iOS example tries to
execute the mkspecs/features/uikit/devices.py python script to find
a simulator ID.

When installing the mkspecs, we didn't preserve the source file
permissions which means that the executable bit was stripped.

This causes the makefile to call xcodebuild with an invalid simulator
id due to not being able to execute the python script.

Make sure to install the mkspecs directories preserving the source
file permissions.

Fixes: QTBUG-87639
Change-Id: Iaa33b740c93187155025891104e8a21c6e7c98c8
Reviewed-by: Cristian Adam <cristian.adam@qt.io>